+++ libs/nest/src/lib/compliance/repositories/compliance-definitions/compliance-definitions.repository.ts
@@ -5,18 +5,45 @@ import { CacheService } from '../../../cache';
 import { compliance_definitions } from '@prisma/client';
 import { set } from 'lodash';
 import { ComplianceResponsesRepository } from '../compliance-responses';
+import { DarklyService } from '../../../darkly';
 
 @Injectable()
 export class ComplianceDefinitionsRepository extends BaseWorker {
-  constructor(readonly prisma: PrismaService, readonly cache: CacheService, readonly responses: ComplianceResponsesRepository) {
+  constructor(readonly prisma: PrismaService, readonly cache: CacheService, readonly responses: ComplianceResponsesRepository, readonly darkly: DarklyService) {
     super(ComplianceDefinitionsRepository.name);
   }
 
-  async getComplianceDefinitionsByOrgId(req: any) {
+  async getComplianceDefinitions() {
     const definitions = (await this.prisma.extended.compliance_definitions.findMany({
-      where: {
-        visible: true,
+      orderBy: {
+        order: 'asc',
       },
+      select: {
+        id: true,
+        name: true,
+        logo_url: true,
+        title: true,
+        visible: true,
+        image_url: true,
+        metadata: true,
+        order: true,
+        version: true,
+      },
+    })) as compliance_definitions[];
+
+    return definitions;
+  }
+
+  async getComplianceDefinitionsByOrgId(req: any, bpc: boolean) {
+    const start = new Date();
+    const list = await this.cache.get(`organization:${req.organization.id}:compliance:list`);
+
+    if (list) {
+      if (!bpc) {
+        return list;
+      }
+    }
+    const definitions = (await this.prisma.extended.compliance_definitions.findMany({
       orderBy: {
         order: 'asc',
       },
@@ -35,27 +62,34 @@ export class ComplianceDefinitionsRepository extends BaseWorker {
 
     if (Array.isArray(definitions) && definitions.length > 0) {
       for (const def of definitions) {
-        const metrics = (await this.cache.get(`organization:${req.organization.id}:compliance:${def.name}:counts`)) as { counts: { progress: number } };
-        if (metrics) {
-          set(def, 'progress', metrics.counts.progress);
-        } else {
-          try {
-            const raw = await this.responses.getScoredComplianceQuestionsByName(req.organization, def.name, req.user);
-            if (raw.counts) {
-              set(def, 'progress', raw.counts.progress);
-            }
-          } catch (e) {
-            if (!(e instanceof NotFoundException)) {
-              this.logger.error(e);
-            }
+        let raw;
+        try {
+          raw = await this.responses.getScoredComplianceQuestionsByName(req.organization, def.name, req.user, undefined, bpc);
+        } catch (e) {
+          if (!(e instanceof NotFoundException)) {
+            this.logger.error(e);
           }
+          continue;
+        }
+
+        // show hidden definitions if their org_compliance is visible
+        if (!def.visible && raw.visible == true) {
+          def.visible = true;
+        }
+
+        if (raw.counts) {
+          set(def, 'progress', raw.counts.progress);
         }
       }
-    } else {
-      throw new NotFoundException('No compliance definitions found');
     }
 
-    return definitions;
+    const end = new Date();
+    this.logger.info(`getComplianceDefinitionsByOrgId completed for ${req.organization.name} in ${end.getTime() - start.getTime()}ms`);
+
+    const filtered = definitions.filter(def => def.visible);
+    await this.cache.set(`organization:${req.organization.id}:compliance:list`, filtered);
+
+    return filtered;
   }
 
   async getComplianceSectionsByComplianceName(name: string) {